+++ index.js
@@ -1,7 +1,7 @@
 'use strict';
 
-import Player from './src/Player';
+import Player, { PlaybackCategories } from './src/Player';
 import Recorder from './src/Recorder';
 import MediaStates from './src/MediaStates';
 
-export { Player, Recorder, MediaStates };
+export { Player, Recorder, MediaStates, PlaybackCategories };